Refactored to extend off a class that allows optional linking
Some content may not need to be directly linked to a category, product or custom URL
so refactored to give us the option and based the new Custom callout on that.

Added a couple of editors to allow switching between plain text and WYSIWYG editor or
force WYSIWYG editor and no plain text.